+++ CHANGES.rst
@@ -59,6 +59,13 @@ Node
   nodes. Also, improved the consistency of ``snapshot`` import errors
   messages (MR :gl:`!10138`)
 
+- Introduced a new process, forked by the node, that is responsible of
+  managing the RPC server: the RPC-process. It is used by default by
+  the node.
+
+- Introduced a new ``--local-rpc-addr`` that starts the RPC server
+  locally, not using the dedicated RPC-process.
+
 Client
 ------
 
@@ -89,6 +96,10 @@ Codec
 Docker Images
 -------------
 
+- The rollup node is protocol agnostic and released as part of the Docker
+  image. (MR :gl:`!10086`)
+
+
 Smart Rollup node
 -----------------
 
@@ -101,6 +112,13 @@ Smart Rollup node
   field ``"log_kernel_debug_file"`` which allows to specify a file in which
   kernel logs should be written (this file is in
   ``<data-dir>/simulation_kernel_logs``). (MR :gl:`!9606`)
+
+- The protocol specific rollup nodes binaries are now deprecated and replaced
+  by symbolic links to the protocol agnostic rollup node. In the future, the
+  symbolic links will be removed. (MR :gl:`!10086`)
+
+- Released the protocol agnostic rollup node ``octez-smart-rollup-node`` as part
+  of the Octez distribution. (MR :gl:`!10086`)
 
 Smart Rollup client
 -------------------