+++ src/kernel_evm/CHANGES.md
@@ -12,6 +12,8 @@
 
 ### Bug fixes
 
+- The internal kernel version is now automatically updated at each upgrade. (!10321)
+
 ### Breaking changes
 
 ### Internal
@@ -54,6 +56,8 @@
 - Add support for `eth_getUncleCountByBlockHash` and `eth_getUncleCountByBlockNumber`. (!10131)
 - Add support for `eth_getUncleByBlockHashAndIndex` and `eth_getUncleByBlockNumberAndIndex`. (!10134)
 - Add 2300 unit of gas for the RPC `eth_estimateGas` to account for the EIP2200. (!10176)
+- Add support for `eth_gasPrice`. (!10195)
+- Add support for `eth_getStorageAt`. (!10214)
 
 ### Bug fixes